backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
yfinance>=0.2.32
pandas>=2.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.10.0
python-multipart>=0.0.6
supabase>=2.13.0
groq>=0.3.0
google-generativeai>=0.3.0
PyPDF2>=3.0.1
pytesseract
Pillow
